add check of result eigenvectors and eigenvalues
- Similarity check functions are found in vector_check.cpp:
-- They account for the fact that the vectors may have
different signs and that the jacobi method has a different order.
I will see if I can add a test that the order of eigvec and eigval
are the same (same eigvec corresponding to same eigval for both).
I have checked manually that this is correct for the jacobi method.
- I did reformulate some things in main
- Changes in ilode are minor print changes